refactor(api): remove membership-repository.js from shared
