+++ api/src/quest/domain/usecases/can-manage-combined-course.js
@@ -6,6 +6,7 @@ export async function canManageCombinedCourse({
 }) {
   const { organizationId } = await combinedCourseRepository.getById({ id: combinedCourseId });
 
+  // bounded-context: should be called with an api from team
   const memberships = await membershipRepository.findByUserIdAndOrganizationId({
     userId,
     organizationId,
