+++ api/src/shared/application/usecases/checkUserBelongsToOrganization.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 
-import * as membershipRepository from '../../infrastructure/repositories/membership-repository.js';
+import * as membershipRepository from '../../../team/infrastructure/repositories/membership.repository.js';
 
 const execute = function (userId, organizationId, dependencies = { membershipRepository }) {
   return dependencies.membershipRepository
