+++ api/src/shared/application/usecases/checkUserBelongsToSupOrganizationAndManagesStudents.js
@@ -1,4 +1,4 @@
-import * as membershipRepository from '../../infrastructure/repositories/membership-repository.js';
+import * as membershipRepository from '../../../team/infrastructure/repositories/membership.repository.js';
 
 const execute = async function (userId, organizationId, dependencies = { membershipRepository }) {
   const memberships = await dependencies.membershipRepository.findByUserIdAndOrganizationId({
